feat: salvar pokemons favoritos em localstorage
